refactor history components
